Match gm_801604DC and gm_80160564 in gm_1601
Both index s16 arrays (lbl_803B7978 etc.) and pass the result to
Toy_8030813C, whose first parameter is s16. The local was declared s32,
forcing a redundant extsh to narrow the value at the call. Declaring it
s16 (matching the array element and parameter types) drops the extsh and
matches.

+++ src/melee/gm/gm_1601.c
@@ -511,7 +511,7 @@ bool gm_80160474(CharacterKind ckind, GameModeKind scene)
 
 char* gm_801604DC(CharacterKind ckind, GameModeKind scene)
 {
-    s32 var_r3;
+    s16 var_r3;
 
     switch (scene) {
     case GM_CLASSIC_GOVER:
@@ -531,7 +531,7 @@ char* gm_801604DC(CharacterKind ckind, GameModeKind scene)
 
 char* gm_80160564(CharacterKind ckind, GameModeKind scene)
 {
-    s32 var_r3;
+    s16 var_r3;
 
     switch (scene) {
     case GM_CLASSIC_GOVER:
